Move `gen_parse_function` to the codegen module
