+++ lrpar/src/lib/codegen.rs
@@ -13,21 +13,22 @@ use std::{
 use crate::{
     LexerTypes, RecoveryKind, RustEdition, SerialisationFormat, Visibility,
     ctbuilder::{
-        ACTION_PREFIX, CTConflictsError, ERROR, FixIntConfig, GLOBAL_PREFIX, VarIntConfig, indent,
-        make_generics,
+        ACTION_PREFIX, ACTIONS_KIND, ACTIONS_KIND_PREFIX, CTConflictsError, ERROR, FixIntConfig,
+        GLOBAL_PREFIX, VarIntConfig, indent, make_generics,
     },
     diagnostics::{DiagnosticFormatter, SpannedDiagnosticFormatter},
 };
 
 use cfgrammar::{
-    Location, Span, Symbol,
+    Location, RIdx, Span, Symbol,
     header::{GrmtoolsSectionParser, Header, HeaderValue},
-    yacc::{YaccGrammar, YaccKind, ast::ASTWithValidityInfo},
+    yacc::{YaccGrammar, YaccKind, YaccOriginalActionKind, ast::ASTWithValidityInfo},
 };
 
 use lrtable::{Minimiser, StateGraph, StateTable, from_yacc};
 use proc_macro2::{Literal, TokenStream};
 use quote::{ToTokens, TokenStreamExt, format_ident, quote};
+use syn::Generics;
 use wincode::SchemaWrite;
 
 pub(crate) struct ParserSrcEnv<'a> {
@@ -712,6 +713,196 @@ where
                 #const_epp_ident[usize::from(tidx)]
             }
         })
+    }
+
+    /// Generate the main parse() function for the output file.
+    pub(crate) fn gen_parse_function(
+        &self,
+        build_env: &ParserBuildEnv<LexerTypesT>,
+    ) -> Result<TokenStream, Box<dyn Error>> {
+        let stable = self.stable();
+        let grm = self.grm();
+        let storaget = str::parse::<TokenStream>(type_name::<LexerTypesT::StorageT>())?;
+        let lexertypest = str::parse::<TokenStream>(type_name::<LexerTypesT>())?;
+        let recoverer = build_env.recoverer();
+        let run_parser = match build_env.yacc_kind() {
+            YaccKind::Original(YaccOriginalActionKind::GenericParseTree) => {
+                quote! {
+                    ::lrpar::RTParserBuilder::new(grm, stable)
+                        .recoverer(#recoverer)
+                        .parse_map(
+                            lexer,
+                            &|lexeme| Node::Term{lexeme},
+                            &|ridx, nodes| Node::Nonterm{ridx, nodes}
+                        )
+                }
+            }
+            YaccKind::Original(YaccOriginalActionKind::NoAction) => {
+                quote! {
+                    ::lrpar::RTParserBuilder::new(grm, stable)
+                        .recoverer(#recoverer)
+                        .parse_map(lexer, &|_| (), &|_, _| ()).1
+                }
+            }
+            YaccKind::Original(YaccOriginalActionKind::UserAction) | YaccKind::Grmtools => {
+                let actionskind = str::parse::<TokenStream>(ACTIONS_KIND)?;
+                let parsed_parse_generics = make_generics(grm.parse_generics().as_deref())?;
+                let (_, type_generics, _) = parsed_parse_generics.split_for_impl();
+                // actions always have a parse_param argument, and when the `parse` function lacks one
+                // that parameter will be unit.
+                let (action_fn_parse_param, action_fn_parse_param_ty) = match grm.parse_param() {
+                    Some((name, ty)) => {
+                        let name = str::parse::<TokenStream>(name)?;
+                        let ty = str::parse::<TokenStream>(ty)?;
+                        (quote!(#name), quote!(#ty))
+                    }
+                    None => (quote!(()), quote!(())),
+                };
+                let wrappers = grm.iter_pidxs().map(|pidx| {
+                    let pidx = usize::from(pidx);
+                    format_ident!("{}wrapper_{}", ACTION_PREFIX, pidx)
+                });
+                let edition_lifetime = if build_env.rust_edition() != RustEdition::Rust2015 {
+                    quote!('_,)
+                } else {
+                    quote!()
+                };
+                let ridx = usize::from(self.user_start_ridx());
+                let action_ident = format_ident!("{}{}", ACTIONS_KIND_PREFIX, ridx);
+
+                quote! {
+                    let actions: ::std::vec::Vec<
+                            &dyn Fn(
+                                    ::cfgrammar::RIdx<#storaget>,
+                                    &'lexer dyn ::lrpar::NonStreamingLexer<'input, #lexertypest>,
+                                    ::cfgrammar::Span,
+                                    ::std::vec::Drain<#edition_lifetime ::lrpar::parser::AStackType<<#lexertypest as ::lrpar::LexerTypes>::LexemeT, #actionskind #type_generics>>,
+                                    #action_fn_parse_param_ty
+                            ) -> #actionskind #type_generics
+                        > = ::std::vec![#(&#wrappers,)*];
+                    match ::lrpar::RTParserBuilder::new(grm, stable)
+                        .recoverer(#recoverer)
+                        .parse_actions(lexer, &actions, #action_fn_parse_param) {
+                            (Some(#actionskind::#action_ident(x)), y) => (Some(x), y),
+                            (None, y) => (None, y),
+                            _ => unreachable!()
+                    }
+                }
+            }
+            kind => panic!("YaccKind {:?} not supported", kind),
+        };
+
+        let parsed_parse_generics: Generics = match build_env.yacc_kind() {
+            YaccKind::Original(YaccOriginalActionKind::UserAction) | YaccKind::Grmtools => {
+                make_generics(grm.parse_generics().as_deref())?
+            }
+            _ => make_generics(None)?,
+        };
+        let (generics, _, where_clause) = parsed_parse_generics.split_for_impl();
+
+        // `parse()` may or may not have an argument for `%parseparam`.
+        let parse_fn_parse_param = match build_env.yacc_kind() {
+            YaccKind::Original(YaccOriginalActionKind::UserAction) | YaccKind::Grmtools => {
+                if let Some((name, tyname)) = grm.parse_param() {
+                    let name = str::parse::<TokenStream>(name)?;
+                    let tyname = str::parse::<TokenStream>(tyname)?;
+                    Some(quote! {#name: #tyname})
+                } else {
+                    None
+                }
+            }
+            _ => None,
+        };
+        let parse_fn_return_ty = match build_env.yacc_kind() {
+            YaccKind::Original(YaccOriginalActionKind::UserAction) | YaccKind::Grmtools => {
+                let actiont = grm
+                    .actiontype(self.user_start_ridx())
+                    .as_ref()
+                    .map(|at| str::parse::<TokenStream>(at))
+                    .transpose()?;
+                quote! {
+                    (::std::option::Option<#actiont>, ::std::vec::Vec<::lrpar::LexParseError<#storaget, #lexertypest>>)
+                }
+            }
+            YaccKind::Original(YaccOriginalActionKind::GenericParseTree) => quote! {
+                (::std::option::Option<Node<<#lexertypest as ::lrpar::LexerTypes>::LexemeT, #storaget>>,
+                    ::std::vec::Vec<::lrpar::LexParseError<#storaget, #lexertypest>>)
+            },
+            YaccKind::Original(YaccOriginalActionKind::NoAction) => quote! {
+                ::std::vec::Vec<::lrpar::LexParseError<#storaget, #lexertypest>>
+            },
+            _ => unreachable!(),
+        };
+
+        let serialisation_format = build_env.serialisation_format();
+        // Note that the configuration types use associated consts, and thus these configurations represent distinct types.
+        let (grm_data, stable_data): (Vec<u8>, Vec<u8>) = match serialisation_format {
+            SerialisationFormat::FixedSizeInteger => {
+                let config = wincode::config::Configuration::default().with_fixint_encoding();
+                let grm = wincode::config::serialize(grm, config)?;
+                let stable = wincode::config::serialize(stable, config)?;
+                (grm, stable)
+            }
+            SerialisationFormat::VariableSizedInteger => {
+                let config = wincode::config::Configuration::default().with_varint_encoding();
+                let grm = wincode::config::serialize(grm, config)?;
+                let stable = wincode::config::serialize(stable, config)?;
+                (grm, stable)
+            }
+        };
+        let serialisation_format_str = quote!(serialisation_format).to_string();
+        Ok(quote! {
+            const __GRM_DATA: &[u8] = &[#(#grm_data,)*];
+            const __STABLE_DATA: &[u8] = &[#(#stable_data,)*];
+            const __SERIALISATION_FORMAT: ::lrpar::ctbuilder::SerialisationFormat = #serialisation_format;
+
+            fn __lrpar_parser_data() -> &'static ::lrpar::ParserData<#storaget> {
+                static DATA: ::std::sync::OnceLock<::lrpar::ParserData<#storaget>>
+                    = ::std::sync::OnceLock::new();
+                DATA.get_or_init(
+                    || {
+                        // We have to call reconstitute like this because the config parameter takes a trait
+                        // which uses const generics. Thus the two config parameters here are not actually of the same type.
+                        match __SERIALISATION_FORMAT {
+                            ::lrpar::ctbuilder::SerialisationFormat::FixedSizeInteger => {
+                                ::lrpar::ctbuilder::_reconstitute(__GRM_DATA, __STABLE_DATA, ::lrpar::ctbuilder::wincode::config::Configuration::default().with_fixint_encoding())
+                            }
+                            ::lrpar::ctbuilder::SerialisationFormat::VariableSizedInteger => {
+                                ::lrpar::ctbuilder::_reconstitute(__GRM_DATA, __STABLE_DATA, ::lrpar::ctbuilder::wincode::config::Configuration::default().with_varint_encoding())
+                            }
+                            _ => {
+                                panic!("Parser source was generated using unknown `SerialisationFormat`: {:?}", #serialisation_format_str)
+                            }
+                        }
+                    }
+                )
+            }
+
+            #[allow(dead_code)]
+            pub fn parse #generics (
+                 lexer: &'lexer dyn ::lrpar::NonStreamingLexer<'input, #lexertypest>,
+                 #parse_fn_parse_param
+            ) -> #parse_fn_return_ty
+            #where_clause
+            {
+                let __data = __lrpar_parser_data();
+                let grm = __data.grm();
+                let stable = __data.stable();
+                #run_parser
+            }
+        })
+    }
+
+    /// Return the `RIdx` of the %start rule in the grammar (which will not be the same as
+    /// grm.start_rule_idx because the latter has an additional rule insert by cfgrammar
+    /// which then calls the user's %start rule).
+    fn user_start_ridx(&self) -> RIdx<LexerTypesT::StorageT> {
+        let grm = self.grm();
+        debug_assert_eq!(grm.prod(grm.start_prod()).len(), 1);
+        match grm.prod(grm.start_prod())[0] {
+            Symbol::Rule(ridx) => ridx,
+            _ => unreachable!(),
+        }
     }
 }
